llm/requirements.txt
chromadb~=0.6.3
openai
flask
flask-cors
requests
google-genai